% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sc_helper_functions.R
\name{ComputeAssessments Computes assessment metrics}
\alias{ComputeAssessments Computes assessment metrics}
\title{Compute Assessments}
\usage{
ComputeAssessments(obj, variables, ...)
}
\description{
This function computes various assessments (LISI, Silhouette, ARI) for multiple
variables in a BenchmarkMetrics object.
}
