% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastRUVIII.R
\name{fastRUVIII}
\alias{fastRUVIII}
\alias{fastRUVIII,Matrix-method}
\alias{fastRUVIII,Seurat-method}
\alias{fastRUVIII,SingleCellExperiment-method}
\title{Run fastRUVIII to remove unwanted variation}
\usage{
fastRUVIII(
  object,
  k,
  replicates = NULL,
  control_genes = NULL,
  assay = NULL,
  apply_log = TRUE
)

\S4method{fastRUVIII}{Matrix}(
  object,
  k,
  replicates = NULL,
  control_genes = NULL,
  assay = NULL,
  apply_log = TRUE
)

\S4method{fastRUVIII}{Seurat}(
  object,
  k,
  replicates = NULL,
  control_genes = NULL,
  assay = NULL,
  apply_log = TRUE
)

\S4method{fastRUVIII}{SingleCellExperiment}(
  object,
  k,
  replicates = NULL,
  control_genes = NULL,
  assay = NULL,
  apply_log = TRUE
)
}
\arguments{
\item{object}{A matrix, Seurat object, or SingleCellExperiment object containing gene expression data}

\item{k}{Integer indicating how much unwanted variation to remove. Lower k = remove less, high k = remove more}

\item{replicates}{A matrix of pseudo replicates, obtained using the CreatePRPC function}

\item{control_genes}{A logical vector indicating which genes are control genes}

\item{assay}{String specifying which assay to use (for Seurat and SingleCellExperiment objects)}

\item{apply_log}{Logical, whether to apply log transformation to the data}
}
\value{
An adjusted expression matrix or an updated Seurat/SingleCellExperiment object with the
        corrected expression data in a new assay called "RUVIII"
}
\description{
This function runs a fast implementation of RUVIII (Remove Unwanted Variation) on
various input types including matrices, Seurat objects, and SingleCellExperiment objects.
}
