% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastRUVIII_Matrix.R
\name{fastRUVIII_Matrix}
\alias{fastRUVIII_Matrix}
\title{Run RUVIII for a data matrix}
\usage{
fastRUVIII_Matrix(matrix, prpc, ncgs, k)
}
\arguments{
\item{matrix}{A matrix with rows as features/genes and columns as cells.}

\item{prpc}{A matrix of your pseudo replicates, obtained using the CreatePRPC function.}

\item{ncgs}{A logical indicating which genes are control genes}

\item{k}{How much unwanted variation to remove. Lower k = remove less, high k = remove more.}
}
\value{
An adjusted (log transformed) matrix with cells as columns and features as rows.
}
\description{
Runs RUVIII and outputs an adjusted matrix
}
