% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindCorrectedMultimodalNeighbours.R
\name{FindCorrectedMultimodalNeighbours}
\alias{FindCorrectedMultimodalNeighbours}
\title{Perform integrative unsupervised clustering for making PRPCs without known biological labels}
\usage{
FindCorrectedMultimodalNeighbours(
  seurat_obj,
  assays,
  uv_variables,
  npcs,
  graph.name = "harmony_wsnn",
  normalization_methods = c("LogNormalize", "CLR")
)
}
\arguments{
\item{seurat_obj}{A seurat object with multi-modal data (each as an assay)}

\item{assays}{A character indicating the names of the assays to use.}

\item{uv_variables}{A character indicating the names of the unwanted variables you do not want to be affected by in your clustering}

\item{npcs}{A numeric vector indicating how many PCs to use for clustering for each modality. Must match length and order of assays.}

\item{graph.name}{Name of the integrated wnn graph to save as.}

\item{normalization_methods}{Names of the quick and fast normalizations to run before clustering. Must match length and order of assays.}
}
\value{
A Seurat object with the integrated shared neighbour graph. Clustering can be done on this graph separately to tune for resolution.
}
\description{
given a seurat object with multi-modal data (each as an assay), do integrative unsupervised clustering
}
