% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindNCG.R
\name{FindNCG}
\alias{FindNCG}
\title{Find negative control genes}
\usage{
FindNCG(
  object,
  unwanted_variables,
  bio_variables,
  assay = NULL,
  metadata = NULL,
  no.ncg = 1000,
  apply_log = T,
  sample_fraction = 0.1
)
}
\arguments{
\item{object}{A dgCMatrix, Seurat or SingleCellExperiment}

\item{unwanted_variables}{A character indicating the names of the unwanted variables
in your metadata that you want to remove.}

\item{bio_variables}{A character indicating the names of the biological groups in your metadata.
RUVIII will assume the variation in these groups is biology and will not be removed.
If you do not already know some previous biology in your data,
labels from unsupervised clustering can also be used. If using multi-omic data,
the FindCorrectedMultimodalNeighbours function could be used for joint multi-omic
clustering (more accurate)}

\item{assay}{The assay containing your RAW COUNTS if using Seurat or SingleCellExperiment}

\item{metadata}{A DATAFRAME of metadata containing the bio and uv variables
(rows as cells and columns as variables)}

\item{no.ncg}{The number of negative control genes to select}

\item{apply_log}{Whether to log transform. TRUE by default.}

\item{sample_fraction}{How much to sample from your data
(for faster computation when using large datasets). Default is 0.1}
}
\value{
A character vector of gene names that are identified as negative control genes.
}
\description{
Given a dgCMatrix, Seurat or SingleCellExperiment, use correlation
analysis to find features/genes that highly associate with unwanted variation
but does not highly associate with biology. (RUVIII will use these genes to
estimate the amount of unwanted variation in the data)
}
