% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sc_helper_functions.R
\name{PlotHVG_Conservation,BenchmarkMetrics-method}
\alias{PlotHVG_Conservation,BenchmarkMetrics-method}
\title{Plot HVG conservation for BenchmarkMetrics objects}
\usage{
\S4method{PlotHVG_Conservation}{BenchmarkMetrics}(
  obj,
  batch_variable = "batch",
  flavour,
  title = "HVG conservation plot",
  n_top_hvgs = 2000
)
}
\description{
Plot HVG conservation for BenchmarkMetrics objects
}
