% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sc_helper_functions.R
\name{PlotHVG_Conservation}
\alias{PlotHVG_Conservation}
\title{Plot HVG conservation generic}
\usage{
PlotHVG_Conservation(
  obj,
  batch_variable = "batch",
  flavour,
  title = "HVG conservation plot",
  n_top_hvgs = 2000
)
}
\description{
Plot HVG conservation generic
}
