% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sc_helper_functions.R
\name{PlotMultipleCorrelations}
\alias{PlotMultipleCorrelations}
\title{Plot multiple correlation plots in a panel}
\usage{
PlotMultipleCorrelations(
  obj,
  variables,
  reductions = "all",
  num_pcs = 10,
  titles = NULL
)
}
\description{
Plot multiple correlation plots in a panel
}
