% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sc_helper_functions.R
\name{RunRUVIII}
\alias{RunRUVIII}
\title{Run RUVIII for benchmarkmetrics object}
\usage{
RunRUVIII(obj, prpc, ncgs, k, name = "RUVIII", pcs = 30)
}
\arguments{
\item{obj}{A BenchmarkMetrics object with populated raw_data and metadata}

\item{prpc}{A matrix of your pseudo replicates}

\item{ncgs}{A logical indicating which genes are control genes}

\item{pcs}{Number of PCs to calculate for the adjusted data}

\item{celltype}{A vector indicating the biological groups}

\item{M}{Replicate matrix}

\item{pseudo_count}{The number to add to the raw counts before log transforming}
}
\value{
A BenchmarkMetrics object with the adjusted data, their PCs and runtimes added to the corresponding slots.
}
\description{
Runs RUVIII, the PCA for its adjusted data, and records runtime.
 All stored back into the BenchmarkMetrics object.
}
