% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{time_eval}
\alias{time_eval}
\title{Runs an expression, times it, and reports the time taken}
\usage{
time_eval(expr, description)
}
\arguments{
\item{expr}{An expression to run}

\item{description}{A character scalar, describing the operation as a noun}
}
\description{
Runs an expression, times it, and reports the time taken
}
